// frontend/src/app/api/generated/model/guest-summary.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { GuestRegistrationType } from './guest-registration-type';


export interface GuestSummary { 
    id: string;
    fullName: string;
    registrationType: GuestRegistrationType;
    createdAt: string;
    activeGymPass: boolean;
    checkedInToday: boolean;
    currentPassName?: string | null;
    currentPassValidUntil?: string | null;
}