// frontend/src/app/api/generated/model/sauna-reservation.ts
/**
 * Excalibur Gym API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { SaunaReservationStatus } from './sauna-reservation-status';


export interface SaunaReservation { 
    id: string;
    guestId: string;
    guestName: string;
    startAt: string;
    endAt: string;
    partySize: number;
    status: SaunaReservationStatus;
    createdAt: string;
    createdByEmployeeName: string;
    cancelledAt?: string | null;
    cancelledByEmployeeName?: string | null;
    cancellationReason?: string | null;
}